Fixed water level type. Added methods to header file
- Measurement from `hc_src04_takeMeasurement` returns a `uint16_t`
- Added the rest of smart pot methods into header file since public methods cannot use private methods

+++ src/smart_pot.c
@@ -33,7 +33,7 @@ uint32_t tryWater() {
 uint8_t getWaterLevel() {
   // 30 + 50 mm
 
-  uint8_t waterLevel = hc_sr04_takeMeasurement() - 30;
+  uint16_t waterLevel = hc_sr04_takeMeasurement() - 30;
   uint8_t waterLevelPercentage = waterLevel / 50 * 100;
 
   return waterLevelPercentage;
